# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing connectors and executing flows step-by-step.",
        "title": "Extended Flow Management System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/executions/{uuid}": {
            "get": {
                "description": "Get full details of an execution including data and steps history",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Get execution details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Execution"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/executions/{uuid}/retry": {
            "post": {
                "description": "Retry a failed execution from the current step, optionally updating input data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Retry a failed execution",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New Input Data to merge",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/executions/{uuid}/step": {
            "get": {
                "description": "Get the definition and JSON schema for the current step. Use this to render the form on the frontend.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Get current step definition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Step"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Validate and submit data for the current step. If successful, advances the flow to the next step.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Submit step data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Execution UUID",
                        "name": "uuid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Form Data matching the step schema",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SubmitStepRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/flows": {
            "get": {
                "description": "Get a list of all available flow definitions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flow-management"
                ],
                "summary": "List all flows",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Flow"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Define a new flow with its steps, schema, and transitions. The 'definition' field must follow the FlowDefinition structure.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "flow-management"
                ],
                "summary": "Create a new Flow Definition",
                "parameters": [
                    {
                        "description": "Flow Definition",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateFlowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Flow"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/flows/{id}/start": {
            "post": {
                "description": "Start a new instance of a flow. Returns the Execution ID (UUID) needed for subsequent steps.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Start a flow execution",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Flow ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Initial input parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StartFlowRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/domain.Execution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "domain.Execution": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "current_step": {
                    "type": "string"
                },
                "data": {
                    "description": "Accumulated data",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "flow_id": {
                    "type": "integer"
                },
                "id": {
                    "description": "UUID",
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/domain.ExecutionStatus"
                },
                "steps_data": {
                    "description": "Map[StepID]Output",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.ExecutionStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "RUNNING",
                "WAITING",
                "COMPLETED",
                "FAILED",
                "SUSPENDED"
            ],
            "x-enum-comments": {
                "StatusSuspended": "Waiting for async callback",
                "StatusWaiting": "Waiting for user input"
            },
            "x-enum-descriptions": [
                "",
                "",
                "Waiting for user input",
                "",
                "",
                "Waiting for async callback"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusRunning",
                "StatusWaiting",
                "StatusCompleted",
                "StatusFailed",
                "StatusSuspended"
            ]
        },
        "domain.Flow": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "definition": {
                    "description": "Stores FlowDefinition",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "next_step": "step2}}",
                        "start_step": "step1",
                        "steps": "{step1:{id:step1",
                        "type": "FORM"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "domain.FlowDefinition": {
            "type": "object",
            "properties": {
                "start_step": {
                    "type": "string",
                    "example": "personal_info"
                },
                "steps": {
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/domain.Step"
                    }
                }
            }
        },
        "domain.Step": {
            "type": "object",
            "properties": {
                "config": {
                    "description": "Connector/Action config",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "param1": "value1"
                    }
                },
                "connector_id": {
                    "description": "For ACTION steps",
                    "type": "integer",
                    "example": 1
                },
                "description": {
                    "type": "string",
                    "example": "Collect personal details"
                },
                "id": {
                    "type": "string",
                    "example": "personal_info"
                },
                "input_mapping": {
                    "description": "Field mapping",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "param1": "{{steps.step1.data.param1}}"
                    }
                },
                "next_step": {
                    "description": "Simple transition",
                    "type": "string",
                    "example": "next_step_id"
                },
                "schema": {
                    "description": "JSON Schema for FORM input",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "properties": "{name:{type:string}}",
                        "required": "[name]",
                        "type": "object"
                    }
                },
                "transitions": {
                    "description": "Conditional transitions",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Transition"
                    }
                },
                "type": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.StepType"
                        }
                    ],
                    "example": "FORM"
                }
            }
        },
        "domain.StepType": {
            "type": "string",
            "enum": [
                "FORM",
                "ACTION",
                "DECISION"
            ],
            "x-enum-comments": {
                "StepTypeAction": "Automated task/connector",
                "StepTypeDecision": "Conditional logic",
                "StepTypeForm": "Human input"
            },
            "x-enum-descriptions": [
                "Human input",
                "Automated task/connector",
                "Conditional logic"
            ],
            "x-enum-varnames": [
                "StepTypeForm",
                "StepTypeAction",
                "StepTypeDecision"
            ]
        },
        "domain.Transition": {
            "type": "object",
            "properties": {
                "condition": {
                    "description": "Simple expression or key",
                    "type": "string",
                    "example": "age \u003e= 18"
                },
                "next_step": {
                    "type": "string",
                    "example": "adult_step"
                }
            }
        },
...
        "dto.CreateFlowRequest": {
            "type": "object",
            "required": [
                "definition",
                "name"
            ],
            "properties": {
                "definition": {
                    "$ref": "#/definitions/domain.FlowDefinition"
                },
                "description": {
                    "type": "string",
                    "example": "Employee onboarding flow with document verification"
                },
                "name": {
                    "type": "string",
                    "example": "Onboarding Process"
                }
            }
        },
...
        "dto.StartFlowRequest": {
            "type": "object",
            "properties": {
                "input": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "campaign_id": "summer_2024",
                        "referrer": "web_portal"
                    }
                }
            }
        },
        "dto.SubmitStepRequest": {
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "age": "30",
                        "email": "john@example.com",
                        "name": "John Doe"
                    }
                }
            }
        }
    }
}